requirements.txt
spotipy>=2.23.0
requests>=2.31.0
lyricsgenius>=3.0.1
python-dotenv>=1.0.0

appdirs
